Use strict int types

+++ binary_search.h
@@ -1,13 +1,16 @@
 #ifndef __BINARY_SEARCH__
 #define __BINARY_SEARCH__
 
+#include <stdint.h>
+
 /**
  * Binary search implementation, searches for `query` in `arr`
  * @param  arr  The array in which the `query` is searched for
  * @param  size The size of the array `arr`
  * @param  query  The element to search for in the array `arr`
- * @return      Returns index of `arr` at which `query` if exists is found else -1
+ * @return      Returns index of `arr` at which `query` is found
+ *              if exists, else -1
  */
-int binary_search(int *arr, int size, int query);
+int16_t binary_search(int_fast64_t *arr, uint8_t size, int_fast64_t query);
 
 #endif
